[v0.1.1]Vectorized input images
Vectoried input images to be in a single matrix and removed loops for iterations in individual matrix. Substantial increase in performance.
Still need to write test functions.
Accuracy seems to come different for this vectorized and non-vectorized vesions. Need to check it if there is something wrong with this implementation or the previous one.

Also corrected calculate_loss_multiple_images function to use the activation function of the network. It was hardcoded to use ReLU.