[v0.1.2]Completed fully connected multi layered multi nodes setup
At start, cost was decreasing but accuracy was not improving. All the outputs were zeros.
I added plots for cost in epochs and observed that there were some great spikes of costs randomly in some epoch. I guessed maybe it is because of high amount of weight update in some epoch.
I put a limit on weight update for 1 and 0.1. and it accuracy improved. But I had to increase the number of epochs for it to learn on training set (5000 epochs)
Test set results have not been fouond impressive.

I also have updated documentation and added some elaborate comments in main functions. Still comments are missing in some small functions.