[v0.1.3]Derivative not giving error
I was inputting numpy object in it. But had to input tf.variable object so that it could get tracked in graph of tensorflow.
According to documentation of gradient tape:

"Operations are recorded if they are executed within this context manager and at least one of their inputs is being "watched".
Trainable variables (created by tf.Variable or tf.get_variable, where trainable=True is default in both cases) are automatically watched. Tensors can be manually watched by invoking the watch method on this context manager."